Break credentials module into parts
