+++ tests/unit_test_cases/verifiable_credential_1_1_from_vc_jwt.json
@@ -19,14 +19,5 @@
   "test_claim_mismatch_issuer",
   "test_claim_mismatch_subject",
   "test_claim_misconfigured_exp",
-  "test_claim_mismatch_exp",
-  "test_validate_dm_empty_id",
-  "test_validate_dm_empty_context",
-  "test_validate_dm_context_base_mismatch",
-  "test_validate_dm_empty_type",
-  "test_validate_dm_type_base_mismatch",
-  "test_validate_dm_empty_issuer",
-  "test_validate_dm_empty_subject",
-  "test_validate_dm_issuance_in_future",
-  "test_validate_dm_credential_expired"
+  "test_claim_mismatch_exp"
 ]
